backend/requirements.txt
flask
flask-cors
pymongo
pandas
openpyxl
python-dotenv